feat: change new user Homepage design to video layout
