--- src/components/Persona/index.js
@@ -1,1 +1,0 @@
-export { PersonaSignup } from './PersonaSignup';
