+++ src/i18n/translations/en/translation.json
@@ -28,67 +28,6 @@
     "p2": "If you're ready to jump in, search or browse the docs. Or create a New Relic account for free. Need some inspiration first? Check out this intro video. Let's begin!"
   },
   "home": {
-    "personas": {
-      "default": {
-        "1": "Front-end devs",
-        "2": "Back-end devs",
-        "3": "Dev-ops and SREs"
-      },
-      "active": {
-        "1": "Front-end developers",
-        "2": "Back-end developers",
-        "3": "Dev-ops and SREs"
-      },
-      "blurbs": {
-        "1": "Click the tiles below to learn how you can use New Relic to monitor the performance of your application, troubleshoot front-end issues your users face, and monitor your site's core web vitals.",
-        "2": "Click the tiles below to learn how you can use New Relic to monitor the performance of your infrastructure, monitor your application, and manage all your logs.",
-        "3": "Click the tiles below to learn how you can use New Relic to monitor your infrastructure, track your changes and deployments, and define service levels between teams."
-      },
-      "tiles": {
-        "1": [
-          {
-            "label": "Monitor your apps",
-            "url": "https://docs.newrelic.com/introduction-apm/"
-          },
-          {
-            "label": "Monitor user experience",
-            "url": "https://docs.newrelic.com/docs/browser/browser-monitoring/getting-started/introduction-browser-monitoring/"
-          },
-          {
-            "label": "Measure core web vitals",
-            "url": "https://docs.newrelic.com/docs/browser/new-relic-browser/guides/guide-to-monitoring-core-web-vitals/"
-          }
-        ],
-        "2": [
-          {
-            "label": "Monitor your infrastructure",
-            "url": "https://docs.newrelic.com/docs/infrastructure/infrastructure-monitoring/get-started/get-started-infrastructure-monitoring/"
-          },
-          {
-            "label": "Monitor your apps",
-            "url": "https://docs.newrelic.com/introduction-apm/"
-          },
-          {
-            "label": "Manage your logs",
-            "url": "https://docs.newrelic.com/docs/logs/get-started/get-started-log-management/"
-          }
-        ],
-        "3": [
-          {
-            "label": "Monitor your infrastructure",
-            "url": "https://docs.newrelic.com/docs/infrastructure/infrastructure-monitoring/get-started/get-started-infrastructure-monitoring/"
-          },
-          {
-            "label": "Track changes and deployments",
-            "url": "https://docs.newrelic.com/docs/change-tracking/change-tracking-introduction/"
-          },
-          {
-            "label": "Define SLIs and SLOs",
-            "url": "https://docs.newrelic.com/docs/service-level-management/create-slm/"
-          }
-        ]
-      }
-    },
     "banner": {
       "title": "Getting started",
       "intro": {
