+++ src/i18n/translations/kr/translation.json
@@ -28,67 +28,6 @@
     "p2": "시작할 준비가 되셨다면 지금 문서를 둘러보거나 무료로 뉴렐릭 계정을 생성하십시오. 먼저 개괄적인 정보를 살펴보길 원하시나요? 지금 이 소개 비디오를 먼저 확인해 보십시오."
   },
   "home": {
-    "personas": {
-      "default": {
-        "1": "프론트엔드 개발자",
-        "2": "백엔드 개발자",
-        "3": "데브옵스 및 SRE"
-      },
-      "active": {
-        "1": "프론트엔드 개발자",
-        "2": "백엔드 개발자",
-        "3": "데브옵스 및 SRE"
-      },
-      "blurbs": {
-        "1": "아래 타일을 클릭하면 뉴렐릭을 사용해 애플리케이션의 성능을 모니터링하고, 사용자가 직면하는 프런트 엔드 문제를 해결하며, 사이트의 코어 웹 바이탈을 모니터링하는 방법을 확인하실 수 있습니다.",
-        "2": "아래 타일을 클릭하면 뉴렐릭을 사용해 인프라의 성능과 애플리케이션을 모니터링하고, 모든 로그를 효과적으로 관리하는 방법을 확인하실 수 있습니다.",
-        "3": "아래 타일을 클릭하면 뉴렐릭을 사용하여 인프라를 모니터링하고, 변경 사항과 배포를 추적하며, 팀 간에 서비스 수준을 정의하는 방법을 확인하실 수 있습니다."
-      },
-      "tiles": {
-        "1": [
-          {
-            "label": "앱 모니터링",
-            "url": "https://docs.newrelic.com/introduction-apm/"
-          },
-          {
-            "label": "사용자 경험 모니터링",
-            "url": "https://docs.newrelic.com/docs/browser/browser-monitoring/getting-started/introduction-browser-monitoring/"
-          },
-          {
-            "label": "코어 웹 바이탈 측정",
-            "url": "https://docs.newrelic.com/docs/browser/new-relic-browser/guides/guide-to-monitoring-core-web-vitals/"
-          }
-        ],
-        "2": [
-          {
-            "label": "인프라 모니터링",
-            "url": "https://docs.newrelic.com/docs/infrastructure/infrastructure-monitoring/get-started/get-started-infrastructure-monitoring/"
-          },
-          {
-            "label": "앱 모니터링",
-            "url": "https://docs.newrelic.com/introduction-apm/"
-          },
-          {
-            "label": "로그 관리",
-            "url": "https://docs.newrelic.com/docs/logs/get-started/get-started-log-management/"
-          }
-        ],
-        "3": [
-          {
-            "label": "인프라 모니터링",
-            "url": "https://docs.newrelic.com/docs/infrastructure/infrastructure-monitoring/get-started/get-started-infrastructure-monitoring/"
-          },
-          {
-            "label": "변경 및 배포 모니터링",
-            "url": "https://docs.newrelic.com/docs/change-tracking/change-tracking-introduction/"
-          },
-          {
-            "label": "SLI 및 SLO 정의",
-            "url": "https://docs.newrelic.com/docs/service-level-management/create-slm/"
-          }
-        ]
-      }
-    },
     "banner": {
       "title": "시작하기",
       "intro": {
